Removed search field from form in Header.js, created SearchBar.js in components/Search to contain all Search functions. Added Form Select dropdown to search. Modified Login/RegisterDropdown.js to resolve render errors.
