+++ frontend/src/components/Auth/RegisterDropdown.js
@@ -64,7 +64,7 @@ const RegisterDropdown = () => {
             color="primary"
             colorVariant="light"
             right="true"
-            icon={<Icon><FontAwesomeIcon icon={faHatWizard} /></Icon>}
+            icon={<Icon align="left"><FontAwesomeIcon icon={faHatWizard} /></Icon>}
             label="Register"
         >
             <Dropdown.Item
