+++ frontend/src/components/Header.js
@@ -5,14 +5,15 @@ import { SearchComicContext } from "../providers/SearchComicProvider";
 import { Navbar, Icon, Form, Button, Container } from "react-bulma-components";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faBookOpen, faCog, faArrowCircleRight, faSearch } from '@fortawesome/free-solid-svg-icons'
-// import { faArrowAltCircleRight,  } from "@fortawesome/free-regular-svg-icons";
 import LoginDropdown from "./Auth/LoginDropdown";
 import RegisterDropdown from "./Auth/RegisterDropdown";
+import SearchBar from "./Search/SearchBar";
 import "../index.css"
 
 
 const Header = () => {
-    // -- GitHub Issue Ticket # 1 ---
+    // 
+    // ---- GitHub Issue Ticket # 1 ----
     // [Authentication [Ticket #1]](https://github.com/jon-xo/bifrost/issues/2)
     // 
     // - Added isLoggedIn state and logout method from UserAccountProvider to check
@@ -23,22 +24,10 @@ const Header = () => {
     // 
 
 
-    const history = useHistory();
-
     const { isLoggedIn, logout } = useContext(UserAccountContext);
-    
-    const [ queryString, setQueryString ] = useState("");
-    const { searchIssues } = useContext(SearchComicContext);
     const [ showBurger, setShowBurger ] = useState(false);
 
-    const handleSearch = (event) => {
-        event.preventDefault();
-        const encodedQuery = encodeURIComponent(queryString)
-        searchIssues(encodedQuery)
-        .then(() => {
-            history.push("/search/issues")
-        })
-    };
+    const history = useHistory();
 
     // -- GitHub Issue Ticket # 2 ---
     // [Navbar [Ticket #2]](https://github.com/jon-xo/bifrost/issues/3)
@@ -120,21 +109,7 @@ const Header = () => {
 
                 <Navbar.Container align='right' transparent='true' className='nav-controls'>
                     <Navbar.Item active={'true'}>
-                        <form>
-                            <Form.Field kind="addons">
-                                <Form.Control>
-                                    <Form.Input placeholder="Find a post" onChange={(e) => {setQueryString(e.target.value)}}/>
-                                </Form.Control>
-                                <Form.Control>
-                                    <Button color='info' onClick={(e) => {handleSearch(e)}}>
-                                        <Icon className='logo-icon'>
-                                            <FontAwesomeIcon icon={faSearch} />
-                                        </Icon>
-                                        Search
-                                    </Button>
-                                </Form.Control>
-                            </Form.Field>
-                        </form>
+                        <SearchBar />
                     </Navbar.Item>
                 </Navbar.Container>  
                 <Navbar.Container align='right'>
